Refactor the Assets classes
Use inheritance instead of traits and also allow assets to load only on certain conditions (specified by calling a closure set in the `enqueue_condition` property of the asset).
